Renamed hdlcc_server.py to server.py
